+++ src/main/java/com/elmakers/mine/bukkit/action/builtin/DisablePhysicsAction.java
@@ -1,0 +1,46 @@
+package com.elmakers.mine.bukkit.action.builtin;
+
+import com.elmakers.mine.bukkit.action.BaseSpellAction;
+import com.elmakers.mine.bukkit.api.action.CastContext;
+import com.elmakers.mine.bukkit.api.magic.MageController;
+import com.elmakers.mine.bukkit.api.spell.SpellResult;
+import com.elmakers.mine.bukkit.spell.BaseSpell;
+import org.bukkit.configuration.ConfigurationSection;
+
+import java.util.Arrays;
+import java.util.Collection;
+
+public class DisablePhysicsAction extends BaseSpellAction
+{
+    private int duration;
+
+    @Override
+    public void prepare(CastContext context, ConfigurationSection parameters) {
+        super.prepare(context, parameters);
+        duration = parameters.getInt("duration", 1);
+    }
+
+	@Override
+	public SpellResult perform(CastContext context) {
+        MageController controller = context.getController();
+        controller.disablePhysics(duration);
+        return SpellResult.CAST;
+	}
+
+    @Override
+    public void getParameterNames(Collection<String> parameters)
+    {
+        super.getParameterNames(parameters);
+        parameters.add("duration");
+    }
+
+    @Override
+    public void getParameterOptions(Collection<String> examples, String parameterKey)
+    {
+        super.getParameterOptions(examples, parameterKey);
+
+        if (parameterKey.equals("duration")) {
+            examples.addAll(Arrays.asList(BaseSpell.EXAMPLE_DURATIONS));
+        }
+    }
+}