+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/GenericSpell.java
@@ -7,6 +7,7 @@ import com.elmakers.mine.bukkit.utility.Target;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Entity;
 
+@Deprecated
 public class GenericSpell extends UndoableSpell
 {
     private boolean undoDamage = false;