+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/PillarSpell.java
@@ -8,6 +8,7 @@ import com.elmakers.mine.bukkit.api.block.MaterialBrush;
 import com.elmakers.mine.bukkit.api.spell.SpellResult;
 import com.elmakers.mine.bukkit.spell.BrushSpell;
 
+@Deprecated
 public class PillarSpell extends BrushSpell 
 {
 	int MAX_SEARCH_DISTANCE = 255;