+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/TreeSpell.java
@@ -9,6 +9,7 @@ import org.bukkit.configuration.ConfigurationSection;
 import com.elmakers.mine.bukkit.api.spell.SpellResult;
 import com.elmakers.mine.bukkit.spell.TargetingSpell;
 
+@Deprecated
 public class TreeSpell extends TargetingSpell
 {
 	private TreeType defaultTreeType = null;